+++ middleware/swagger-metadata.js
@@ -89,6 +89,7 @@ var expressStylePath = function (basePath, apiPath) {
   // Replace Swagger syntax for path parameters with Express' version (All Swagger path parameters are required)
   return (basePath + apiPath).replace(/{/g, ':').replace(/}/g, '');
 };
+
 var convertValue = function (value, schema, type) {
   var original = value;
 
@@ -104,6 +105,11 @@ var convertValue = function (value, schema, type) {
 
   // If there is no value, do not convert it
   if (_.isUndefined(value)) {
+    return value;
+  }
+
+  // If there is an empty value and allowEmptyValue is true, return it
+  if (schema.allowEmptyValue && value === '') {
     return value;
   }
 
@@ -146,7 +152,11 @@ var convertValue = function (value, schema, type) {
 
   case 'integer':
     if (!_.isNumber(value)) {
-      value = parseInt(value, 10);
+      if (_.isString(value) && _.trim(value).length === 0) {
+        value = NaN;
+      }
+
+      value = Number(value);
 
       if (isNaN(value)) {
         value = original;
@@ -157,7 +167,11 @@ var convertValue = function (value, schema, type) {
 
   case 'number':
     if (!_.isNumber(value)) {
-      value = parseFloat(value);
+      if (_.isString(value) && _.trim(value).length === 0) {
+        value = NaN;
+      }
+
+      value = Number(value);
 
       if (isNaN(value)) {
         value = original;
@@ -181,6 +195,7 @@ var convertValue = function (value, schema, type) {
 
   return value;
 };
+
 var processOperationParameters = function (version, pathKeys, pathMatch, req, res, next) {
   var swaggerMetadata = req.swagger;
   var parameters = !_.isUndefined(swaggerMetadata) ?